move the global ca-bundle handling to existing ca-bundle infrastructure
